+++ host/src/main/kotlin/by/egorr/gradle/ktsexec/host/KtsExecKotlinScriptHost.kt
@@ -2,23 +2,31 @@
 package by.egorr.gradle.ktsexec.host
 
 import java.io.File
+import kotlin.reflect.KClass
 import kotlin.script.experimental.annotations.KotlinScript
 import kotlin.script.experimental.api.EvaluationResult
+import kotlin.script.experimental.api.KotlinType
 import kotlin.script.experimental.api.ResultWithDiagnostics
+import kotlin.script.experimental.host.createCompilationConfigurationFromTemplate
 import kotlin.script.experimental.host.toScriptSource
+import kotlin.script.experimental.jvm.defaultJvmScriptingHostConfiguration
 import kotlin.script.experimental.jvm.dependenciesFromCurrentContext
 import kotlin.script.experimental.jvm.jvm
 import kotlin.script.experimental.jvmhost.BasicJvmScriptingHost
-import kotlin.script.experimental.jvmhost.createJvmCompilationConfigurationFromTemplate
 import kotlin.system.exitProcess
 
+@Suppress("Unused")
 @KotlinScript
 abstract class SimpleScript
 
-inline fun <reified S : Any> evalFile(
+internal fun evalFile(
+    scriptDefinition: KClass<out Any>,
     scriptFile: File
 ): ResultWithDiagnostics<EvaluationResult> {
-    val compilationConfiguration = createJvmCompilationConfigurationFromTemplate<S> {
+    val compilationConfiguration = createCompilationConfigurationFromTemplate(
+        baseClassType = KotlinType(scriptDefinition),
+        hostConfiguration = defaultJvmScriptingHostConfiguration
+    ) {
         jvm {
             dependenciesFromCurrentContext(wholeClasspath = true)
         }
@@ -32,11 +40,17 @@ fun main(vararg args: String) {
         exitProcess(-1)
     }
 
+    println("Host arguments: ${args.joinToString()}")
+
+    val scriptDefinition = Class.forName(args[1].substringAfter('=')).kotlin
+    println("Script definition class: $scriptDefinition")
+
     val scriptFile = File(args[0])
     println("Start script $scriptFile execution.")
 
-    val res = evalFile<SimpleScript>(scriptFile)
+    val res = evalFile(scriptDefinition, scriptFile)
 
+    println("Reports: ${res.reports}")
     res.reports.forEach {
         println(" : ${it.message}" + if (it.exception == null) "" else ": ${it.exception} \n ${it.exception?.printStackTrace()}")
     }