+++ plugin/src/main/kotlin/by/egorr/gradle/plugin/ktsexec/KtsExecPlugin.kt
@@ -1,10 +1,11 @@
 package by.egorr.gradle.plugin.ktsexec
 
+import by.egorr.gradle.ktsexec.KTS_EXEC_CONFIGURATION_NAME
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 
 open class KtsExecPlugin : Plugin<Project> {
     override fun apply(target: Project) {
-        // no-op
+        target.configurations.maybeCreate(KTS_EXEC_CONFIGURATION_NAME)
     }
 }